+++ Pang/Pang/Source/ModuleRender.cpp
@@ -94,7 +94,7 @@ bool ModuleRender::CleanUp()
 }
 
 // Blit to screen
-bool ModuleRender::Blit(SDL_Texture* texture, int x, int y, const SDL_Rect* section, float speed, bool useCamera)
+bool ModuleRender::Blit(SDL_Texture* texture, float x, float y, const SDL_Rect* section, float speed, bool useCamera)
 {
 	bool ret = true;
